feat: find transactions by categoryId and extra validation on category delete

+++ src/services/CategoriesServiceImpl.ts
@@ -1,9 +1,12 @@
 import { Category } from "../domains/Category";
+import { ConflictError } from "../errors/ConflictError";
 import { CategoriesRepository } from "../repositories/CategoriesRepository";
+import { TransactionsRepository } from "../repositories/TransactionsRepository";
 import { CategoriesService } from "./CategoriesService";
 
 export interface CategoriesServiceProps {
   categoriesRepo: CategoriesRepository;
+  transactionsRepo: TransactionsRepository;
 }
 
 export class CategoriesServiceImpl implements CategoriesService {
@@ -27,6 +30,18 @@ export class CategoriesServiceImpl implements CategoriesService {
   }
 
   async delete(category: Category): Promise<void> {
+    const transactions = await this.props.transactionsRepo.findByCategoryId(
+      category.user.id!,
+      category.id,
+    );
+
+    if (transactions.length > 0) {
+      throw new ConflictError({
+        detail:
+          "Category cannot be deleted because it is associated with active transactions.",
+      });
+    }
+
     await this.props.categoriesRepo.delete(category);
   }
 }
